refactor: modularize floating search overlay
