+++ prds/watch/work.md
@@ -1,0 +1,29 @@
+# Search Component Overlay Refactor
+
+## Goals
+- Extract floating search overlay logic into a reusable hook.
+- Split SearchComponent presentation into focused view components.
+- Add automated coverage for overlay focus management and trending fallback behaviour.
+
+## Implementation Strategy
+- [x] Audit current SearchComponent responsibilities and identify logic to move into a hook.
+- [x] Create `useFloatingSearchOverlay` hook to manage search state, focus handling, and trending fetch lifecycle.
+- [x] Extract `CategoryGrid`, `QuickList`, and overlay result layouts into standalone components in `SearchComponent/`.
+- [x] Update `SearchComponent.tsx` to consume the hook and render the new presentational pieces, keeping the file under 150 lines.
+- [x] Author unit tests for the hook (focus/closing) and components (trending fallback, layout rendering).
+- [x] Run lint/test suite relevant to the watch app and ensure clean git status.
+
+## Risks & Mitigations
+- **Radix dialog focus trapping** could swallow blur events. Mitigate by unit tests that exercise blur handlers through the hook.
+- **Trending API failures** must fall back to translations. Cover with tests that simulate hook error state.
+- Ensure Algolia refine resets correctly when overlay closes via Escape or blur.
+
+## Validation Steps
+- Overlay opens on input focus and auto-focuses the search field.
+- Quick select chips trigger new search and keep overlay open.
+- Trending list falls back to translated defaults on fetch error.
+- Tests for new hook/components pass.
+
+## Follow-up Ideas
+- Consider moving category metadata to CMS-driven config if design requires frequent updates.
+- Evaluate migrating remaining MUI layout primitives to Tailwind equivalents in a future iteration.
